feat: add label-has-for rule
Adds a new rule for jsx-a11y/label-has-for.
This rule was fixed in a recent version of jsx-a11y and is now flagging incorrectly in the Reaction repo. Adding this rule resolves the linting errors.

This rule configuration was recommended in this stack-overflow Q&A https://stackoverflow.com/questions/44951806/jsx-a11y-returning-form-label-must-have-associated-control-when-there-is-an-html

The label-has-for rule is now deprecated and will be replaced by label-has-associated-control
https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/label-has-for.md